fix: cookie payment 패키지 경로 정리
