--- src/main/java/com/domisa/domisa_backend/domain/payment/dto/CreateCookieOrderRequest.java
@@ -1,7 +1,0 @@
-package com.domisa.domisa_backend.domain.payment.dto;
-
-public record CreateCookieOrderRequest(
-	Integer cookieAmount,
-	Integer orderAmount
-) {
-}
