--- src/main/java/com/domisa/domisa_backend/domain/payment/dto/CreateCookieOrderResponse.java
@@ -1,13 +1,0 @@
-package com.domisa.domisa_backend.domain.payment.dto;
-
-public record CreateCookieOrderResponse(
-	String orderNumber,
-	Integer orderAmount,
-	String billingName,
-	String bankName,
-	String bankCode,
-	String accountNumber,
-	String accountHolder,
-	String status
-) {
-}
