--- src/main/java/com/domisa/domisa_backend/domain/payment/dto/PayActionCreateOrderResponse.java
@@ -1,10 +1,0 @@
-package com.domisa.domisa_backend.domain.payment.dto;
-
-public record PayActionCreateOrderResponse(
-	String status,
-	PayActionResponseBody response
-) {
-	public boolean isSuccess() {
-		return "success".equalsIgnoreCase(status);
-	}
-}
