--- src/main/java/com/domisa/domisa_backend/domain/payment/dto/PayActionMatchedWebhookRequest.java
@@ -1,10 +1,0 @@
-package com.domisa.domisa_backend.domain.payment.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public record PayActionMatchedWebhookRequest(
-	@JsonProperty("order_number") String orderNumber,
-	@JsonProperty("order_status") String orderStatus,
-	@JsonProperty("processing_date") String processingDate
-) {
-}
