--- src/main/java/com/domisa/domisa_backend/domain/payment/dto/PayActionResponseBody.java
@@ -1,6 +1,0 @@
-package com.domisa.domisa_backend.domain.payment.dto;
-
-public record PayActionResponseBody(
-	String message
-) {
-}
