--- src/main/java/com/domisa/domisa_backend/domain/payment/dto/PayActionWebhookResponse.java
@@ -1,9 +1,0 @@
-package com.domisa.domisa_backend.domain.payment.dto;
-
-public record PayActionWebhookResponse(
-	String status
-) {
-	public static PayActionWebhookResponse success() {
-		return new PayActionWebhookResponse("success");
-	}
-}
