--- src/main/java/com/domisa/domisa_backend/domain/payment/entity/CookieTransactionType.java
@@ -1,7 +1,0 @@
-package com.domisa.domisa_backend.domain.payment.entity;
-
-public enum CookieTransactionType {
-	CHARGE,
-	USE,
-	REFUND
-}
