--- src/main/java/com/domisa/domisa_backend/domain/payment/entity/OrderStatus.java
@@ -1,8 +1,0 @@
-package com.domisa.domisa_backend.domain.payment.entity;
-
-public enum OrderStatus {
-	PENDING,
-	PAID,
-	CANCELED,
-	FAILED
-}
