+++ src/main/java/com/domisa/domisa_backend/payment/controller/CookieOrderController.java
@@ -1,0 +1,28 @@
+package com.domisa.domisa_backend.payment.controller;
+
+import com.domisa.domisa_backend.auth.annotation.AuthUser;
+import com.domisa.domisa_backend.payment.dto.CreateCookieOrderRequest;
+import com.domisa.domisa_backend.payment.dto.CreateCookieOrderResponse;
+import com.domisa.domisa_backend.payment.service.CookieOrderService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/orders/cookies")
+@RequiredArgsConstructor
+public class CookieOrderController {
+
+	private final CookieOrderService cookieOrderService;
+
+	@PostMapping
+	public ResponseEntity<CreateCookieOrderResponse> createCookieOrder(
+		@AuthUser Long userId,
+		@RequestBody CreateCookieOrderRequest request
+	) {
+		return ResponseEntity.ok(cookieOrderService.createCookieOrder(userId, request));
+	}
+}
