+++ src/main/java/com/domisa/domisa_backend/payment/dto/CreateCookieOrderRequest.java
@@ -1,0 +1,7 @@
+package com.domisa.domisa_backend.payment.dto;
+
+public record CreateCookieOrderRequest(
+	Integer cookieAmount,
+	Integer orderAmount
+) {
+}
