+++ src/main/java/com/domisa/domisa_backend/payment/dto/PayActionCreateOrderResponse.java
@@ -1,0 +1,10 @@
+package com.domisa.domisa_backend.payment.dto;
+
+public record PayActionCreateOrderResponse(
+	String status,
+	PayActionResponseBody response
+) {
+	public boolean isSuccess() {
+		return "success".equalsIgnoreCase(status);
+	}
+}
