+++ src/main/java/com/domisa/domisa_backend/payment/dto/PayActionResponseBody.java
@@ -1,0 +1,6 @@
+package com.domisa.domisa_backend.payment.dto;
+
+public record PayActionResponseBody(
+	String message
+) {
+}
