+++ src/main/java/com/domisa/domisa_backend/payment/dto/PayActionWebhookResponse.java
@@ -1,0 +1,9 @@
+package com.domisa.domisa_backend.payment.dto;
+
+public record PayActionWebhookResponse(
+	String status
+) {
+	public static PayActionWebhookResponse success() {
+		return new PayActionWebhookResponse("success");
+	}
+}
