+++ src/main/java/com/domisa/domisa_backend/payment/entity/CookieTransactionType.java
@@ -1,0 +1,7 @@
+package com.domisa.domisa_backend.payment.entity;
+
+public enum CookieTransactionType {
+	CHARGE,
+	USE,
+	REFUND
+}
