+++ src/main/java/com/domisa/domisa_backend/payment/entity/OrderStatus.java
@@ -1,0 +1,8 @@
+package com.domisa.domisa_backend.payment.entity;
+
+public enum OrderStatus {
+	PENDING,
+	PAID,
+	CANCELED,
+	FAILED
+}
